Add javadoc comment and clean up code.
