+++ cas-server-webapp/src/main/resources/messages.properties
@@ -53,7 +53,7 @@ screen.rememberme.checkbox.title=Remember Me
 
 # Blocked Errors Page
 screen.blocked.header=Access Denied
-screen.blocked.message=You've been throttled for too frequent login attempts.
+screen.blocked.message=You've been throttled for suspicious login attempts. Please wait for a few seconds before login again.
 AbstractAccessDecisionManager.accessDenied=You are not authorized to access this resource. Contact your CAS administrator for more info.
 
 #Confirmation Screen Messages
