+++ etc/cas.properties
@@ -282,7 +282,7 @@ log4j.config.location=classpath:log4j2.xml
 
 ##
 # Login Throttle
-cas.throttle.failureRangeInSeconds = 1
-cas.throttle.failureThreshold = 3
+cas.throttle.failureRangeInSeconds = 2
+cas.throttle.failureThreshold = 4
 cas.throttle.startDelay = 0
-cas.throttle.repeatInterval = 1000
+cas.throttle.repeatInterval = 500
